[GEARPUMP-226] refactor build files
