+++ streaming/src/main/scala/org/apache/gearpump/streaming/task/Subscription.scala
@@ -20,7 +20,7 @@ package org.apache.gearpump.streaming.task
 
 import org.slf4j.Logger
 
-import org.apache.gearpump.google.common.primitives.Shorts
+import com.google.common.primitives.Shorts
 import org.apache.gearpump.streaming.partitioner.{MulticastPartitioner, Partitioner, UnicastPartitioner}
 import org.apache.gearpump.streaming.AppMasterToExecutor.MsgLostException
 import org.apache.gearpump.streaming.LifeTime
